Fix more warnings and some style.

+++ src/house.c
@@ -219,7 +219,7 @@ static void House_save_control(void)
     return;
   }
   /* write all the house control recs in one fell swoop.  Pretty nifty, eh? */
-  if (fwrite(house_control, sizeof(struct house_control_rec), num_of_houses, fl) != num_of_houses) {
+  if (fwrite(house_control, sizeof(struct house_control_rec), num_of_houses, fl) != (size_t)num_of_houses) {
     perror("SYSERR: Unable to save house control file.");
     return;	  
   }
